+++ src/templates/react-app/react/base/src/App.mustache.js
@@ -3,12 +3,14 @@ import { Home } from './components/Home';
 {{#models}}
 import { {{modelNamePascal}}Router } from './components/{{modelNamePascal}}/{{modelNamePascal}}Router';
 {{/models}}
+import { Header } from "./components/App/Header";
 import './App.scss';
 
 function App() {
   return (
     <Router>
       <div className="App">
+        <Header />
         <Switch>
         {{#models}}
           <Route path="/{{modelNamePascal}}">