+++ src/templates/react-app/react/base/src/components/App/Drawer.mustache.js
@@ -1,0 +1,53 @@
+import { Link } from 'react-router-dom';
+import {
+  Drawer as MaterialDrawer,
+  List,
+  ListItem,
+  ListItemText,
+  ListItemIcon,
+  IconButton,
+  Divider
+} from '@material-ui/core';
+import { ChevronLeft as ChevronLeftIcon, Home as HomeIcon } from '@material-ui/icons';
+import './Drawer.scss';
+
+export function Drawer(props) {
+  const { open, handleClose } = props;
+
+  return (
+    <div className="Drawer">
+      <MaterialDrawer
+        className="drawer"
+        variant="persistent"
+        anchor="left"
+        open={open}
+        classes={{=<% %>=}}{{
+          paper: 'drawerPaper'
+        }}<%={{ }}=%>
+      >
+        <div className="drawerHeader">
+          <IconButton onClick={handleClose}>
+            <ChevronLeftIcon />
+          </IconButton>
+        </div>
+        <Divider />
+        <List>
+          <ListItem button key={'Home'} component={Link} to="/">
+            <ListItemIcon>
+              <HomeIcon />
+            </ListItemIcon>
+            <ListItemText primary={'Home'} />
+          </ListItem>
+        </List>
+        <Divider />
+        <List>
+          {{#models}}
+          <ListItem button component={Link} to="/{{modelNamePascal}}">
+            <ListItemText primary="{{modelName}}" />
+          </ListItem>
+          {{/models}}
+        </List>
+      </MaterialDrawer>
+    </div>
+  );
+}